# apps/web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Tempo API",
    "version": "0.5.0"
  },
  "paths": {
...
    "/api/corrections/recalculate-prs": {
      "post": {
        "tags": [
          "corrections"
        ],
        "summary": "Recalculate Prs",
        "description": "Rebuild personal records and their chronology from the live sets + hand-entered claims.",
        "operationId": "recalculate_prs_api_corrections_recalculate_prs_post",
        "parameters": [
          {
            "name": "exercise_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Exercise Id"
            }
          },
          {
            "name": "pr_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Pr Type"
            }
          },
          {
            "name": "dry_run",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Dry Run"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecalculationOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/corrections/verify-prs": {
      "get": {
        "tags": [
          "corrections"
        ],
        "summary": "Verify Prs",
        "description": "Read-only: report any exercise whose record tables disagree with themselves.",
        "operationId": "verify_prs_api_corrections_verify_prs_get",
        "parameters": [
          {
            "name": "exercise_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Exercise Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IntegrityOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/corrections/restore": {
      "post": {
        "tags": [
          "corrections"
        ],
        "summary": "Restore",
        "description": "Undo a soft delete and rebuild whatever records depended on the row being gone.",
        "operationId": "restore_api_corrections_restore_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RestoreIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestoreOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/corrections/purge": {
      "post": {
        "tags": [
          "corrections"
        ],
        "summary": "Purge",
        "description": "Hard-delete rows soft-deleted longer ago than ``older_than_days``. Irreversible.",
        "operationId": "purge_api_corrections_purge_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PurgeIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PurgeOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "IntegrityOut": {
        "properties": {
          "ok": {
            "type": "boolean",
            "title": "Ok"
          },
          "checked_exercises": {
            "type": "integer",
            "title": "Checked Exercises"
          },
          "problems": {
            "items": {
              "$ref": "#/components/schemas/IntegrityProblemOut"
            },
            "type": "array",
            "title": "Problems"
          }
        },
        "type": "object",
        "required": [
          "ok",
          "checked_exercises",
          "problems"
        ],
        "title": "IntegrityOut"
      },
      "IntegrityProblemOut": {
        "properties": {
          "exercise_id": {
            "type": "string",
            "format": "uuid",
            "title": "Exercise Id"
          },
          "exercise_name": {
            "type": "string",
            "title": "Exercise Name"
          },
          "pr_type": {
            "type": "string",
            "title": "Pr Type"
          },
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "detail": {
            "type": "string",
            "title": "Detail"
          }
        },
        "type": "object",
        "required": [
          "exercise_id",
          "exercise_name",
          "pr_type",
          "kind",
          "detail"
        ],
        "title": "IntegrityProblemOut"
      },
...
      "MetricResultOut": {
        "properties": {
          "exercise_id": {
            "type": "string",
            "format": "uuid",
            "title": "Exercise Id"
          },
          "exercise_name": {
            "type": "string",
            "title": "Exercise Name"
          },
          "pr_type": {
            "type": "string",
            "title": "Pr Type"
          },
          "value": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Value"
          },
          "source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source"
          },
          "entries": {
            "type": "integer",
            "title": "Entries"
          },
          "changed": {
            "type": "boolean",
            "title": "Changed"
          }
        },
        "type": "object",
        "required": [
          "exercise_id",
          "exercise_name",
          "pr_type",
          "value",
          "source",
          "entries",
          "changed"
        ],
        "title": "MetricResultOut",
        "description": "What a rebuild produced for one (exercise, metric)."
      },
...
      "PurgeIn": {
        "properties": {
          "older_than_days": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Older Than Days"
          },
          "dry_run": {
            "type": "boolean",
            "title": "Dry Run",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "older_than_days"
        ],
        "title": "PurgeIn"
      },
      "PurgeOut": {
        "properties": {
          "older_than_days": {
            "type": "integer",
            "title": "Older Than Days"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "counts": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Counts"
          },
          "dry_run": {
            "type": "boolean",
            "title": "Dry Run"
          }
        },
        "type": "object",
        "required": [
          "older_than_days",
          "total",
          "counts",
          "dry_run"
        ],
        "title": "PurgeOut"
      },
      "RecalculationOut": {
        "properties": {
          "exercises": {
            "type": "integer",
            "title": "Exercises"
          },
          "changed": {
            "items": {
              "$ref": "#/components/schemas/MetricResultOut"
            },
            "type": "array",
            "title": "Changed"
          },
          "metrics": {
            "items": {
              "$ref": "#/components/schemas/MetricResultOut"
            },
            "type": "array",
            "title": "Metrics"
          },
          "dry_run": {
            "type": "boolean",
            "title": "Dry Run",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "exercises",
          "changed",
          "metrics"
        ],
        "title": "RecalculationOut"
      },
      "RestoreIn": {
        "properties": {
          "entity_type": {
            "type": "string",
            "title": "Entity Type"
          },
          "entity_id": {
            "type": "string",
            "format": "uuid",
            "title": "Entity Id"
          }
        },
        "type": "object",
        "required": [
          "entity_type",
          "entity_id"
        ],
        "title": "RestoreIn"
      },
      "RestoreOut": {
        "properties": {
          "entity_type": {
            "type": "string",
            "title": "Entity Type"
          },
          "entity_id": {
            "type": "string",
            "format": "uuid",
            "title": "Entity Id"
          },
          "exercises_recalculated": {
            "type": "integer",
            "title": "Exercises Recalculated"
          }
        },
        "type": "object",
        "required": [
          "entity_type",
          "entity_id",
          "exercises_recalculated"
        ],
        "title": "RestoreOut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}